remove customer provider
